refactor: consolidate edge-aware wavelets code (#7224)
* move eaw_decompose and eaw_synthesize to new file common/eaw.c

* move eaw_decompose and eaw_synthesize from denoiseprofile.c to common/eaw.c

* add more restrict qualifiers, ensure alignment of params

validated against tests 0013 and 0024

* apply row-interleaving for better cache behavior

(for denoiseprofile only; the current atrous version of the code is
not compatible with the interleaving trick)

Thr  Old    New    (test 0013)
4    2.894   2.830  (-2.2%)
8    1.595   1.466  (-8%)
16   0.906   0.794  (-12%)
32   0.504   0.459  (-9%)

* add #include needed for Win64
